MapTransform changes
Since the ScriptEngine does not work well with double, it has been
changed to use BigDecimal instead, and move required operations to the
specification of the operator to use, to avoid any extra processing and
parsing that will delay considerably the processing time. This fixes the
double precision problem and reduces the calculations to the basic four
operators, but it can be extended to as many as required.