FileDataSet unit tests
Unit tests for the File type of DataSet which will extend the DataSet
abstract class, implementing required methods and applying core
functionality to it, making the file available and configurable in terms
of training and testing sets, and also at the row level between input
columns (the input vector for the neural network inputs), and the output
columns (the output vector for the neural network outputs). The output
is only used on supervised neural networks.