Changing template and tactics
When loading a trained network, the results found when testing the
accuracy at each 10 epochs, do not match the final result. E.g.: For a
network trained up to 66% accuracy, when loading the saved encog file
with all the network trained data, it is now showing only a 45%
accuracy. When comparing with other networks, it was found a correlation
with the error and time training. Therefore, changing the template to
train indefinitely and saving encog files every 100 epoch until the
accuracy is close to expected value. Removing all tests done