Normalised values stored
When training a neural network, normalised min and max, need to be saved
for next training with a different set. This is to keep the neural
network trained integrity. The min and max must be to min and max for
all possible event ranges, or it will cause problems in the future -
something to think about: how to predict possible future min and max
that may go out of range in the future..